[#432] Update credential form to use most recently active one
