+++ apigee_edge.links.task.yml
@@ -137,8 +137,8 @@ apigee_edge.developer_app.analytics_for_developer:
   base_route: entity.developer_app.canonical_by_developer
   weight: -1
 
-apigee_edge.developer_app.generate_credential:
-  route_name: entity.developer_app.generate_credential_form
-  title: 'Generate credentials'
+apigee_edge.developer_app.add_credential:
+  route_name: entity.developer_app.add_credential_form
+  title: 'Add credentials'
   base_route: entity.developer_app.canonical_by_developer
   weight: -1
