+++ modules/apigee_edge_teams/src/Form/TeamAppCredentialAddForm.php
@@ -17,41 +17,35 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-namespace Drupal\apigee_edge\Form;
+namespace Drupal\apigee_edge_teams\Form;
 
 use Drupal\apigee_edge\Entity\AppInterface;
-use Drupal\apigee_edge\Entity\Form\DeveloperAppFormTrait;
+use Drupal\apigee_edge\Form\AppCredentialAddFormBase;
+use Drupal\apigee_edge_teams\Entity\Form\TeamAppFormTrait;
+use Drupal\apigee_edge_teams\Entity\TeamInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Url;
-use Drupal\user\UserInterface;
 
 /**
- * Provides credential generate form for developer app.
+ * Provides credential add form for team app.
  */
-class DeveloperAppCredentialGenerateForm extends AppCredentialGenerateFormBase {
+class TeamAppCredentialAddForm extends AppCredentialAddFormBase {
 
-  use DeveloperAppFormTrait, DeveloperAppCredentialFormTrait;
+  use TeamAppFormTrait, TeamAppCredentialFormTrait;
 
   /**
-   * The user from route.
+   * The team from route.
    *
-   * @var \Drupal\user\UserInterface
+   * @var \Drupal\apigee_edge_teams\Entity\TeamInterface
    */
-  protected $user;
+  protected $team;
 
   /**
    * {@inheritdoc}
    */
-  public function buildForm(array $form, FormStateInterface $form_state, ?AppInterface $app = NULL, ?UserInterface $user = NULL) {
-    $this->user = $user;
+  public function buildForm(array $form, FormStateInterface $form_state, ?AppInterface $app = NULL, ?TeamInterface $team = NULL) {
+    $this->team = $team;
     return parent::buildForm($form, $form_state, $app);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function getAppOwner(): string {
-    return $this->user->getEmail();
   }
 
   /**
@@ -59,6 +53,13 @@ class DeveloperAppCredentialGenerateForm extends AppCredentialGenerateFormBase {
    */
   protected function getRedirectUrl(): Url {
     return $this->getCancelUrl();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function getAppOwner(): string {
+    return $this->team->id();
   }
 
 }
