+++ src/Entity/DeveloperApp.php
@@ -55,7 +55,7 @@ use Drupal\user\UserInterface;
  *       "delete_for_developer" = "Drupal\apigee_edge\Entity\Form\DeveloperAppDeleteFormForDeveloper",
  *       "analytics" = "Drupal\apigee_edge\Form\DeveloperAppAnalyticsForm",
  *       "analytics_for_developer" = "Drupal\apigee_edge\Form\DeveloperAppAnalyticsFormForDeveloper",
- *       "generate_credential" = "Drupal\apigee_edge\Form\DeveloperAppCredentialGenerateForm",
+ *       "add_credential" = "Drupal\apigee_edge\Form\DeveloperAppCredentialAddForm",
  *       "delete_credential" = "Drupal\apigee_edge\Form\DeveloperAppCredentialDeleteForm",
  *       "revoke_credential" = "Drupal\apigee_edge\Form\DeveloperAppCredentialRevokeForm",
  *     },
@@ -78,7 +78,7 @@ use Drupal\user\UserInterface;
  *     "edit-form-for-developer" = "/user/{user}/apps/{app}/edit",
  *     "delete-form-for-developer" = "/user/{user}/apps/{app}/delete",
  *     "analytics-for-developer" = "/user/{user}/apps/{app}/analytics",
- *     "generate-credential-form" = "/user/{user}/apps/{app}/credentials/generate",
+ *     "add-credential-form" = "/user/{user}/apps/{app}/credentials/add",
  *     "delete-credential-form" = "/user/{user}/apps/{app}/credentials/{consumer_key}/delete",
  *     "revoke-credential-form" = "/user/{user}/apps/{app}/credentials/{consumer_key}/revoke",
  *   },
