+++ src/Entity/DeveloperAppRouteProvider.php
@@ -26,7 +26,7 @@ use Drupal\apigee_edge\Entity\ListBuilder\DeveloperAppListBuilderForDeveloper;
 use Drupal\apigee_edge\Form\DeveloperAppAnalyticsFormForDeveloper;
 use Drupal\apigee_edge\Form\DeveloperAppCredentialApproveForm;
 use Drupal\apigee_edge\Form\DeveloperAppCredentialDeleteForm;
-use Drupal\apigee_edge\Form\DeveloperAppCredentialGenerateForm;
+use Drupal\apigee_edge\Form\DeveloperAppCredentialAddForm;
 use Drupal\apigee_edge\Form\DeveloperAppCredentialRevokeForm;
 use Drupal\Core\Entity\EntityTypeInterface;
 use Symfony\Component\Routing\Route;
@@ -67,8 +67,8 @@ class DeveloperAppRouteProvider extends AppRouteProvider {
       $collection->add("entity.{$entity_type_id}.analytics_for_developer", $analytics_for_developer);
     }
 
-    if ($generate_credential_form = $this->getGenerateCredentialRoute($entity_type)) {
-      $collection->add("entity.{$entity_type_id}.generate_credential_form", $generate_credential_form);
+    if ($add_credential_form = $this->getAddCredentialRoute($entity_type)) {
+      $collection->add("entity.{$entity_type_id}.add_credential_form", $add_credential_form);
     }
 
     if ($delete_credential_form = $this->getDeleteCredentialRoute($entity_type)) {
@@ -240,7 +240,7 @@ class DeveloperAppRouteProvider extends AppRouteProvider {
   }
 
   /**
-   * Gets the generate-credential-form route for a developer app.
+   * Gets the add-credential-form route for a developer app.
    *
    * @param \Drupal\Core\Entity\EntityTypeInterface $entity_type
    *   The entity type.
@@ -248,11 +248,11 @@ class DeveloperAppRouteProvider extends AppRouteProvider {
    * @return \Symfony\Component\Routing\Route|null
    *   The generated route, if available.
    */
-  protected function getGenerateCredentialRoute(EntityTypeInterface $entity_type) {
-    if ($entity_type->hasLinkTemplate('generate-credential-form')) {
-      $route = new Route($entity_type->getLinkTemplate('generate-credential-form'));
-      $route->setDefault('_form', DeveloperAppCredentialGenerateForm::class);
-      $route->setDefault('_title', 'Generate credentials');
+  protected function getAddCredentialRoute(EntityTypeInterface $entity_type) {
+    if ($entity_type->hasLinkTemplate('add-credential-form')) {
+      $route = new Route($entity_type->getLinkTemplate('add-credential-form'));
+      $route->setDefault('_form', DeveloperAppCredentialAddForm::class);
+      $route->setDefault('_title', 'Add credentials');
       $route->setDefault('entity_type_id', $entity_type->id());
       $this->ensureUserParameter($route);
       $route->setRequirement('_app_access_check_by_app_name', 'update');
